refact: move code from backend.__init__.py to files
Also convert OP CODES from the peephole evaluator to
enums.
