+++ src/arch/z80/backend/_array.py
@@ -16,7 +16,7 @@ from src.arch.z80.backend._32bit import _32bit_oper
 from src.arch.z80.backend._f16 import _f16_oper
 from src.arch.z80.backend._float import _float_oper, _fpush
 from src.arch.z80.backend.common import REQUIRES, Quad, is_int, runtime_call
-from src.arch.z80.backend.errors import InvalidICError
+from src.arch.z80.backend.exception import InvalidICError
 from src.arch.z80.backend.runtime import Labels as RuntimeLabel
