+++ tests/api/test_arg_parser.py
@@ -1,5 +1,6 @@
 import unittest
 
+from src.arch.z80 import backend
 from src.zxbc.args_parser import parser
 
 
@@ -7,6 +8,7 @@ class TestArgParser(unittest.TestCase):
     """Test argument options from the cmdline"""
 
     def setUp(self):
+        backend.init()  # backend contains "org" option needed in these tests
         self.parser = parser()
 
     def test_autorun_defaults_to_none(self):
